refactor: 일정에 색상 반영
