+++ src/entities/calendar/model/scheduleStore.tsx
@@ -1,8 +1,11 @@
 import { ScheduleResponse } from '@/entities/calendar'
 import { addDays, isWithinInterval } from 'date-fns'
 import { create } from 'zustand'
+import { COLORS } from '../consts/constants'
 
 interface IScheduleStore {
+  color: string
+  setColor: (color: string) => void
   title: string | null
   note: string | null
   date: { startDateTime: Date; endDateTime: Date } | null
@@ -28,6 +31,8 @@ interface IScheduleStore {
 }
 
 export const useScheduleStore = create<IScheduleStore>((set) => ({
+  color: COLORS.LIGHT_PURPLE,
+  setColor: (color: string) => set({ color }),
   title: null,
   note: null,
   date: null,
