+++ src/views/calendar/ui/AddEvent/AddEvent.tsx
@@ -3,6 +3,7 @@
 import AddEventTitle from './AddScheduleTitle'
 import AddEventTime from './AddScheduleTime'
 import AddEventNote from './AddScheduleNote'
+import AddScheduleColor from './AddScheduleColor'
 import { useParams, useRouter } from 'next/navigation'
 import { useMutation, useQuery } from '@tanstack/react-query'
 import {
@@ -23,6 +24,7 @@ import { BaseButton, BaseHeader, LoadingSpinner } from '@/shared/ui'
 import AddEventLocation from './AddScheduleLocation'
 import { COLORS } from '@/entities/calendar/consts/constants'
 
+
 /**
  * id 있다면 : 스케쥴 수정
  * id 없다면 : 스케쥴 생성
@@ -32,6 +34,8 @@ export function AddEventPage() {
     title,
     note,
     date,
+    color,
+    setColor,
     setTitle,
     setNote,
     setDate,
@@ -99,13 +103,15 @@ export function AddEventPage() {
         startDateTime: new Date(scheduleDetailData.startDateTime),
         endDateTime: new Date(scheduleDetailData.endDateTime),
       })
+      setColor(scheduleDetailData.color)
     }
     return () => {
       setTitle(null)
       setNote(null)
       setDate(null)
+      setColor(COLORS.LIGHT_PURPLE)
     }
-  }, [scheduleDetailData, setTitle, setNote, setDate, scheduleId])
+  }, [scheduleDetailData, setColor, setTitle, setNote, setDate, scheduleId])
 
   /**
    * Schedule 저장
@@ -119,7 +125,7 @@ export function AddEventPage() {
       location: '',
       startDateTime: format(date?.startDateTime || new Date(baseDate), "yyyy-MM-dd'T'HH:mm:ss"),
       endDateTime: format(date?.endDateTime || new Date(baseDate), "yyyy-MM-dd'T'HH:mm:ss"),
-      color: findKeyByValue(color),
+      color: findKeyByValue(color) || 'LIGHT_PURPLE',
     }
 
     //시작시간이 끝나는 시간보다 클 경우
@@ -138,29 +144,12 @@ export function AddEventPage() {
     deleteMutation.mutate()
   }
 
-  // 임시 color 부여 로직
-  const [color, setColor] = useState(COLORS.LIGHT_PURPLE)
 
-  const onClickColor = (color: string) => {
-    handleColorChange(color)
-    setIsColorPickerOpen(false)
-  }
-
-  const handleColorChange = (color: string) => {
-    setColor(color)
-  }
-
-  const [isColorPickerOpen, setIsColorPickerOpen] = useState(false)
-
-  const handleColorPickerToggle = () => {
-    setIsColorPickerOpen(!isColorPickerOpen)
-  }
 
   // color 의 key 찾기
-  const findKeyByValue = (value: string) => {
-    return Object.keys(COLORS).find((key) => COLORS[key] === value)
+  const findKeyByValue = (color: string) => {
+    return Object.keys(COLORS).find((key) => COLORS[key as keyof typeof COLORS] === color)
   }
-
 
   if (isLoading) return <LoadingSpinner />
   if (isError) alert(error)
@@ -186,25 +175,8 @@ export function AddEventPage() {
       <div className="flex flex-col px-[1.5rem] overflow-hidden py-[1.5rem] h-[75vh] ">
         <div className="flex flex-col flex-grow overflow-y-auto gap-[3rem] ">
           <div className='relative flex justify-between items-center'>
-          <AddEventTitle placeholder={'Title'} />
-          <div className='w-[1em] h-[1em] rounded-full cursor-pointer' 
-          style={{ backgroundColor: color }} 
-          onClick={handleColorPickerToggle}
-          />
-          {
-            isColorPickerOpen && (
-              <div className='absolute top-[30px] right-[5px] grid grid-cols-5 p-[10px] gap-[0.5em] rounded-[20px] bg-white shadow-[0_2px_8px_rgba(0,0,0,0.1)]'>
-                {Object.values(COLORS).map((color) => (
-                  <div
-                    key={`plan-color-${color}`}
-                    className='w-[1em] h-[1em] rounded-full cursor-pointer'
-                    style={{ backgroundColor: color }}
-                    onClick={() => onClickColor(color)}
-                  />
-                ))}
-              </div>
-            )
-          }
+            <AddEventTitle placeholder={'Title'} />
+            <AddScheduleColor />
           </div>
           <AddEventTime />
           <AddEventLocation />
